pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "indico-orpheus"
version = "0.1.0"
description = "Reusable automation utilities and workflows for Indico"
requires-python = ">=3.11"
dependencies = [
    "httpx",
    "python-dotenv",
    "PyGithub",
    "pandas",
    "requests",
    "PyJWT",
    "geopy",
    "googlemaps",
    "indico-client"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
